splitting the challenge screen into an active and a completed screen
